wallet.py
#!/usr/bin/env python
# -*- coding: utf-8 -*-

import os
import logging
import secrets
import hashlib
from dotenv import load_dotenv
from web3 import Web3
from cryptography.fernet import Fernet

# Chargement des variables d'environnement
load_dotenv()
BLOCKCHAIN_API_KEY = os.getenv('BLOCKCHAIN_API_KEY')

class Wallet:
    def __init__(self):
        self.logger = logging.getLogger(__name__)
        # Cru00e9er une clu00e9 de chiffrement pour su00e9curiser les clu00e9s privu00e9es
        # Dans une application ru00e9elle, cette clu00e9 devrait u00eatre stocku00e9e en su00e9curitu00e9
        self.encryption_key = Fernet.generate_key() if not os.getenv('ENCRYPTION_KEY') else os.getenv('ENCRYPTION_KEY').encode()
        self.cipher_suite = Fernet(self.encryption_key)
        
        # Initialiser la connexion Web3 (si nu00e9cessaire pour interagir avec une blockchain ru00e9elle)
        # Pour l'instant, nous simulons ces opu00e9rations
        self.w3 = None
        try:
            # Exemple avec Ethereum (Infura)
            if BLOCKCHAIN_API_KEY:
                self.w3 = Web3(Web3.HTTPProvider(f"https://mainnet.infura.io/v3/{BLOCKCHAIN_API_KEY}"))
                self.logger.info("Connexion u00e0 la blockchain u00e9tablie")
        except Exception as e:
            self.logger.warning(f"Impossible de se connecter u00e0 la blockchain: {str(e)}")
            self.logger.info("Fonctionnement en mode simulation")
        
        # Ru00e9fu00e9rence u00e0 la base de donnu00e9es (sera injectu00e9e par le bot)
        self.db = None
    
    def set_db(self, db):
        """Du00e9finir la ru00e9fu00e9rence u00e0 la base de donnu00e9es"""
        self.db = db
    
    def create_wallet(self, user_id):
        """Cru00e9er un nouveau wallet pour un utilisateur"""
        try:
            # Gu00e9nu00e9rer une clu00e9 privu00e9e et une adresse (simulu00e9 pour l'exemple)
            private_key = self._generate_private_key()
            address = self._generate_address_from_private_key(private_key)
            
            # Chiffrer la clu00e9 privu00e9e pour le stockage
            encrypted_private_key = self._encrypt_private_key(private_key)
            
            # Stocker dans la base de donnu00e9es
            if self.db:
                success = self.db.create_wallet(user_id, address, encrypted_private_key.decode())
                if success:
                    self.logger.info(f"Wallet cru00e9u00e9 pour l'utilisateur {user_id}")
                    return address
            return None
        except Exception as e:
            self.logger.error(f"Erreur lors de la cru00e9ation du wallet: {str(e)}")
            return None
    
    def get_wallet_address(self, user_id):
        """Obtenir l'adresse du wallet d'un utilisateur"""
        try:
            if self.db:
                wallet_info = self.db.get_wallet(user_id)
                if wallet_info:
                    return wallet_info['address']
            return None
        except Exception as e:
            self.logger.error(f"Erreur lors de la ru00e9cupu00e9ration de l'adresse du wallet: {str(e)}")
            return None
    
    def get_balance(self, user_id):
        """Obtenir le solde du wallet d'un utilisateur"""
        try:
            if self.db:
                wallet_info = self.db.get_wallet(user_id)
                if wallet_info:
                    # Dans une application ru00e9elle, nous interrogerions la blockchain
                    # Pour l'exemple, nous utilisons le solde stocku00e9 dans la base de donnu00e9es
                    return wallet_info['balance']
            return 0.0
        except Exception as e:
            self.logger.error(f"Erreur lors de la ru00e9cupu00e9ration du solde: {str(e)}")
            return 0.0
    
    def deposit(self, user_id, amount, tx_hash=None):
        """Simuler un du00e9pu00f4t de fonds dans le wallet de l'utilisateur"""
        try:
            if self.db:
                current_balance = self.get_balance(user_id)
                new_balance = current_balance + amount
                success = self.db.update_balance(user_id, new_balance)
                
                if success:
                    # Enregistrer la transaction
                    self.db.record_transaction(
                        sender_id=0,  # 0 pour repru00e9senter un du00e9pu00f4t externe
                        receiver_id=user_id,
                        amount=amount,
                        transaction_type='deposit',
                        status='completed',
                        tx_hash=tx_hash
                    )
                    self.logger.info(f"Du00e9pu00f4t de {amount} effectuu00e9 pour l'utilisateur {user_id}")
                    return f"Du00e9pu00f4t de {amount} CRYPTO effectuu00e9 avec succu00e8s. Nouveau solde: {new_balance} CRYPTO"
            return "Erreur lors du du00e9pu00f4t. Veuillez ru00e9essayer."
        except Exception as e:
            self.logger.error(f"Erreur lors du du00e9pu00f4t: {str(e)}")
            return f"Erreur lors du du00e9pu00f4t: {str(e)}"
    
    def withdraw(self, user_id, to_address, amount):
        """Retirer des fonds du wallet de l'utilisateur"""
        try:
            if self.db:
                current_balance = self.get_balance(user_id)
                
                # Vu00e9rifier si le solde est suffisant
                if current_balance < amount:
                    return "Solde insuffisant pour effectuer ce retrait"
                
                # Dans une application ru00e9elle, nous enverrions les fonds sur la blockchain
                # Pour l'exemple, nous simulons le retrait
                new_balance = current_balance - amount
                success = self.db.update_balance(user_id, new_balance)
                
                if success:
                    # Enregistrer la transaction
                    tx_hash = self._generate_tx_hash()  # Simulu00e9
                    self.db.record_transaction(
                        sender_id=user_id,
                        receiver_id=0,  # 0 pour repru00e9senter un retrait externe
                        amount=amount,
                        transaction_type='withdrawal',
                        status='completed',
                        tx_hash=tx_hash
                    )
                    self.logger.info(f"Retrait de {amount} effectuu00e9 par l'utilisateur {user_id}")
                    return f"Retrait de {amount} CRYPTO effectuu00e9 avec succu00e8s vers {to_address}. Nouveau solde: {new_balance} CRYPTO"
            return "Erreur lors du retrait. Veuillez ru00e9essayer."
        except Exception as e:
            self.logger.error(f"Erreur lors du retrait: {str(e)}")
            return f"Erreur lors du retrait: {str(e)}"
    
    def transfer(self, from_user_id, to_user_id, amount):
        """Transfu00e9rer des fonds entre deux utilisateurs"""
        try:
            if self.db:
                sender_balance = self.get_balance(from_user_id)
                
                # Vu00e9rifier si le solde est suffisant
                if sender_balance < amount:
                    return False, "Solde insuffisant pour effectuer ce transfert"
                
                # Mettre u00e0 jour le solde de l'expu00e9diteur
                new_sender_balance = sender_balance - amount
                self.db.update_balance(from_user_id, new_sender_balance)
                
                # Mettre u00e0 jour le solde du destinataire
                receiver_balance = self.get_balance(to_user_id)
                new_receiver_balance = receiver_balance + amount
                self.db.update_balance(to_user_id, new_receiver_balance)
                
                # Enregistrer la transaction
                tx_hash = self._generate_tx_hash()  # Simulu00e9
                self.db.record_transaction(
                    sender_id=from_user_id,
                    receiver_id=to_user_id,
                    amount=amount,
                    transaction_type='transfer',
                    status='completed',
                    tx_hash=tx_hash
                )
                
                self.logger.info(f"Transfert de {amount} de {from_user_id} vers {to_user_id}")
                return True, f"Transfert de {amount} CRYPTO effectuu00e9 avec succu00e8s. Nouveau solde: {new_sender_balance} CRYPTO"
            return False, "Erreur lors du transfert. Veuillez ru00e9essayer."
        except Exception as e:
            self.logger.error(f"Erreur lors du transfert: {str(e)}")
            return False, f"Erreur lors du transfert: {str(e)}"
    
    # Mu00e9thodes privu00e9es pour les opu00e9rations cryptographiques
    
    def _generate_private_key(self):
        """Gu00e9nu00e9rer une clu00e9 privu00e9e alu00e9atoire (simulu00e9)"""
        # Pour une application ru00e9elle, utilisez les bibliothu00e8ques crypto appropriu00e9es
        return secrets.token_hex(32)
    
    def _generate_address_from_private_key(self, private_key):
        """Gu00e9nu00e9rer une adresse u00e0 partir d'une clu00e9 privu00e9e (simulu00e9)"""
        # Pour une application ru00e9elle, utilisez les bibliothu00e8ques crypto appropriu00e9es
        # Ici, nous gu00e9nu00e9rons simplement une adresse ETH-like
        h = hashlib.sha256(private_key.encode()).hexdigest()
        return f"0x{h[-40:]}"
    
    def _encrypt_private_key(self, private_key):
        """Chiffrer la clu00e9 privu00e9e pour le stockage su00e9curisu00e9"""
        return self.cipher_suite.encrypt(private_key.encode())
    
    def _decrypt_private_key(self, encrypted_private_key):
        """Du00e9chiffrer la clu00e9 privu00e9e"""
        if isinstance(encrypted_private_key, str):
            encrypted_private_key = encrypted_private_key.encode()
        return self.cipher_suite.decrypt(encrypted_private_key).decode()
    
    def _generate_tx_hash(self):
        """Gu00e9nu00e9rer un hash de transaction simulu00e9"""
        return f"0x{secrets.token_hex(32)}"
