requirements.txt
python-telegram-bot==13.15
python-dotenv==1.0.0
sqlite3-api==2.0.4
requests==2.31.0
cryptography==41.0.3
pycryptodome==3.18.0
web3==6.0.0
pytz==2025.2
